+++ crypto/keys/multisig/amino.go
@@ -3,7 +3,7 @@ package multisig
 import (
 	errorsmod "cosmossdk.io/errors"
 
-	types "github.com/cosmos/cosmos-sdk/codec/types"
+	"github.com/cosmos/cosmos-sdk/codec/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )